Refactor part-parsing.  All but 2 tests passing.
This refactors a lot of things, but was tricky to do atomically.

The first part is moving the comma-delimited parsing to a separate
function, along with several other things that were being done
inline unnecessarily.

The next is to avoid a bunch of excessive parsing of range parts,
and the prefix which comes before any matched open-{.